% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi}
\alias{bmab_gi}
\alias{bmab_gi_ab}
\title{Calculate Gittins indices for a single arm (Bernoulli rewards)}
\usage{
bmab_gi(Sigma, n, gamma, N, tol = 5e-04, lb = NA, ub = NA)

bmab_gi_ab(alpha, beta, gamma, N, tol = 5e-04, lb = NA, ub = NA)
}
\arguments{
\item{Sigma}{Value of Sigma for the arm.}

\item{n}{Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{N}{Integer > 0; time horizon used.}

\item{tol}{Absolute accuracy required.}

\item{lb}{Optional lower bound for GI.}

\item{ub}{Optional upper bound for GI.}

\item{alpha}{Value of alpha for the arm.}

\item{beta}{Value of beta for the arm.}
}
\value{
A single Gittins index
}
\description{
There are two versions which differ only in their parameterisation and in the layout of the output.
\code{bmab_gi()} uses \code{Sigma} and \code{n} parameters, while \code{bmab_gi_ab()} uses \code{alpha = Sigma}
and  \code{beta = n - Sigma}.
}
\details{
The initial interval for calibration are as follows:
For lower bound, use lb if supplied else use KGI.
For upper bound, use ub if supplied else use GI+.
}
