% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_ab}
\alias{bmab_gi_ab}
\title{Calculate Gittins indices for a single arm (Bernoulli rewards)}
\usage{
bmab_gi_ab(alpha, beta, gamma, tol, N, lb = NA, ub = NA, kgi = F,
  giplus = F)
}
\arguments{
\item{alpha}{value of alpha for the arm}

\item{beta}{value of beta for the arm}

\item{gamma}{discount factor}

\item{tol}{absolute accuracy required}

\item{N}{time horizon used}

\item{lb=NA}{optional lower bound for GI}

\item{ub=NA}{optional upper bound for GI}

\item{kgi=F}{optional boolean indicates whether to use KGI for lower bound (only if lb=NA)}

\item{giplus=F}{optional boolean indicates whether to use GI+ for upper bound (only if ub=NA)}
}
\value{
A single Gittins index
}
\description{
This version uses alpha, beta parameterisation.
The initial interval for calibration are as follows:
For lower bound, use lb if supplied else use KGI if kgi=T or Sigma/n otherwise.
For upper bound, use ub if supplied else use GI+ if giplus=T or 1 otherwise.
}
