% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_multiple_ab}
\alias{bmab_gi_multiple_ab}
\title{Calculates Gittins indices for the BMAB with a range of arm states.}
\usage{
bmab_gi_multiple_ab(alpha_start = 1, beta_start = 1, gamma, N,
  num_actions, tol)
}
\arguments{
\item{gamma}{discount factor}

\item{N}{time horizon used}

\item{num_actions}{determinines the number of states GI are calculated for}

\item{tol}{absolute accuracy required}

\item{alpha_start=1}{lowest value of alpha for the arms}

\item{beta_start=1}{lowest value of beta for the arms}
}
\value{
A triangular matrix of GI values
}
\description{
This version uses alpha, beta parameterisation.
The states are a triangular matrix with:
alpha=alpha_start, alpha_start + 1, ..., alpha_start + num_actions - 1,
beta=beta_start, beta_start + 1, ..., beta_start + num_actions - 1,
and alpha + beta <= num_actions + alpha_start + beta_start - 1.
}
