% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_giplus}
\alias{bmab_giplus}
\title{Calculates the GI+ index for the BMAB for a single arm with supplied parameters.}
\usage{
bmab_giplus(Sigma, n, gamma, tol, upper = F)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}

\item{tol}{absolute accuracy required}

\item{upper=F}{if TRUE, the upper end of the interval is returned, otherwise the midpoint}
}
\value{
A vector of GI+ values
}
\description{
Upper bound for GI.
}
