% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_giplus_value}
\alias{bmab_giplus_value}
\title{Value approximation for the one-armed bandit with Normal rewards using GI+ as policy.}
\usage{
bmab_giplus_value(lambda, Sigma, n, gamma)
}
\arguments{
\item{lambda}{reward from the known arm}

\item{Sigma}{value of Sigma for the unknown arm}

\item{n}{value of n for the unknown arm}

\item{gamma}{discount factor}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Value approximation for the one-armed bandit with Normal rewards using GI+ as policy.
}
