% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_kgi}
\alias{bmab_kgi}
\title{Calculates the KGI index for the BMAB for a single arm with supplied parameters.}
\usage{
bmab_kgi(Sigma, n, gamma)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}
}
\value{
A vector of KGI values
}
\description{
Exact closed form calculation.
}
