% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calibrate_arm}
\alias{calibrate_arm}
\title{Calibrate an unknown arm against a known arm}
\usage{
calibrate_arm(FUN, lb, ub, tol, ...)
}
\arguments{
\item{FUN}{A value function.}

\item{lb}{Initial lower end of interval.}

\item{ub}{Initial upper end of interval.}

\item{tol}{Absolute accuracy required.}

\item{...}{Other arguments passed to \code{FUN}.}
}
\value{
A vector of lower and upper end of an interval containing the true value.
}
\description{
Performs iterative reduction of an interval using calibration with value function FUN.
}
