% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gittins-package.R
\docType{package}
\name{gittins-package}
\alias{gittins}
\alias{gittins-package}
\title{gittins: Calculate Gittins Indices}
\description{
Calculates Gittins indices for the multi-armed bandit problem.
}
\keyword{internal}
