% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_args}
\alias{nmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{Value of Sigma for the arm.}

\item{n}{Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tau}{Observation precision.}

\item{tol}{Absolute accuracy required.}

\item{N}{Time horizon used.}

\item{xi}{Value of xi (entent of dynamic program state space).}

\item{delta}{Value of delta (fineness of discretisation in the dynamic program).}
}
\description{
Function arguments
}
\keyword{internal}
