% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi_value}
\alias{nmab_gi_value}
\title{Value calculation for the one-armed bandit with Normal rewards.}
\usage{
nmab_gi_value(lambda, n, gamma, tau, N, xi, delta)
}
\arguments{
\item{lambda}{reward from the known arm}

\item{n}{value of n for the unknown arm}

\item{gamma}{discount factor}

\item{tau}{observation precision}

\item{N}{time horizon used}

\item{xi}{value of xi (entent of dynamic programmme state space)}

\item{delta}{value of delta (fineness of discretisation in the dynamic programmme)}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Assumes Sigma=mu=0.
}
