% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_giplus}
\alias{nmab_giplus}
\title{Calculates the GI+ index for the NMAB for a single arm with supplied parameters.}
\usage{
nmab_giplus(Sigma, n, gamma, tol, lb, ub, upper = F)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}

\item{tol}{absolute accuracy required}

\item{upper=F}{if TRUE, the upper end of the interval is returned, otherwise the midpoint}
}
\value{
A vector of GI values
}
\description{
The index is an upper bound for GI.
}
