% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_giplus_value}
\alias{nmab_giplus_value}
\title{Value approximation for the one-armed bandit with Normal rewards using GI+ as policy.}
\usage{
nmab_giplus_value(lambda, mu, n, gamma)
}
\arguments{
\item{lambda}{reward from the known arm}

\item{mu}{mean of reward belief for the unknown arm}

\item{n}{value of n for the unknown arm}

\item{gamma}{discount factor}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Value approximation for the one-armed bandit with Normal rewards using GI+ as policy.
}
