% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_kgi}
\alias{nmab_kgi}
\title{Calculate the knowledge gradient index for a single arm (normal rewards)}
\usage{
nmab_kgi(Sigma, n, gamma, tau, tol, ub = NA, lower = FALSE)
}
\arguments{
\item{Sigma}{value of Sigma for the arm.}

\item{n}{value of n for the arm.}

\item{gamma}{numeric in (0, 1); discount factor.}

\item{tau}{observation precision.}

\item{tol}{absolute accuracy required.}

\item{ub}{Optional upper bound for the index.}

\item{lower}{If TRUE, the lower end of the interval is returned, otherwise the midpoint.}
}
\value{
A vector of GI values.
}
\description{
The index is an lower bound for GI.
}
