% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_kgi_value}
\alias{nmab_kgi_value}
\title{Value of one-armed bandit using KGI (normal rewards)}
\usage{
nmab_kgi_value(lambda, mu, n, gamma, tau)
}
\arguments{
\item{lambda}{reward from the known arm}

\item{mu}{mean of reward belief for the unknown arm}

\item{n}{value of n for the unknown arm}

\item{gamma}{discount factor}

\item{tau}{observation precision}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Calculation is as given in Ryzhov, Powell & Frazier (2012).
}
