% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_risky_reward}
\alias{nmab_risky_reward}
\title{Helper function only used in nmab_gi_value.}
\usage{
nmab_risky_reward(mu, y_lo_scaled, y_hi_scaled, tn_scaled, tau, s,
  value_vec, discount)
}
\description{
Returns the value of the risky arm with the supplied parameters.
}
