% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_v_args}
\alias{nmab_v_args}
\title{Function arguments for value functions}
\arguments{
\item{lambda}{reward from the known arm}

\item{mu}{mean of reward belief for the unknown arm}

\item{n}{value of n for the unknown arm}
}
\description{
Function arguments for value functions
}
\keyword{integral}
