% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_args}
\alias{bmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{numeric in (0, 1); discount factor}

\item{tol}{absolute accuracy required}

\item{N}{integer>0; time horizon used}
}
\description{
Function arguments
}
\keyword{internal}
