% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_args}
\alias{bmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{Numeric > 0. Value of Sigma for the arm.}

\item{n}{Numeric > Sigma > 0. Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tol}{Numeric > 0. Absolute accuracy required.}

\item{N}{Integer >= 2; time horizon used.}
}
\description{
Function arguments
}
\keyword{internal}
