% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_args}
\alias{bmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{Value of Sigma for the arm.}

\item{n}{Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tol}{Absolute accuracy required.}

\item{N}{Integer > 0; time horizon used.}
}
\description{
Function arguments
}
\keyword{internal}
