% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi}
\alias{bmab_gi}
\title{Calculate Gittins indices for a single arm (Bernoulli rewards)}
\usage{
bmab_gi(Sigma, n, gamma, tol, N, lb = NA, ub = NA, kgi = F,
  giplus = F)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{numeric in (0, 1]; discount factor}

\item{tol}{absolute accuracy required}

\item{N}{integer>0; time horizon used}

\item{lb=NA}{optional lower bound for GI}

\item{ub=NA}{optional upper bound for GI}

\item{kgi=F}{optional boolean indicates whether to use KGI for lower bound (only if \code{lb=NA})}

\item{giplus=F}{optional boolean indicates whether to use GI+ for upper bound (only if \code{ub=NA})}
}
\value{
A single Gittins index
}
\description{
There are two versions which differ only in their parameterisation and in the layout of the output.
\code{bmab_gi()} uses \code{Sigma} and \code{n} parameters, while \code{bmab_gi_ab()} uses \code{alpha = Sigma}
and  \code{beta = n - Sigma}.
}
\details{
The initial interval for calibration are as follows:
For lower bound, use lb if supplied else use KGI if \code{kgi=T} or \code{Sigma/n} otherwise.
For upper bound, use ub if supplied else use GI+ if \code{giplus=T} or \code{1} otherwise.
}
