% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_multiple}
\alias{bmab_gi_multiple}
\title{Calculate Gittins indices for multiple starting states (Bernoulli rewards)}
\usage{
bmab_gi_multiple(
  alpha_start = NULL,
  beta_start = NULL,
  gamma,
  N,
  num_actions,
  tol = 0.0005,
  Sigma_start = NULL,
  n_start = NULL
)
}
\arguments{
\item{alpha_start}{Lowest value of alpha for the arms.}

\item{beta_start}{Lowest value of beta for the arms.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{N}{Integer > 0; time horizon used.}

\item{num_actions}{Determines the number of states GI are calculated for.}

\item{tol}{Absolute accuracy required.}

\item{Sigma_start}{Lowest value of Sigma for the arms.}

\item{n_start}{Lowest value of n for the arms.}
}
\value{
A data frame of starting states and GI values. Extra information is stored in attributes:
\itemize{
\item \code{params}: the parameters used.
\item \code{gi_matrix}: the GI values in a matrix (beta x alpha).
\item \code{gi_matrix_ns}: the GI values in a matrix (n x Sigma).
}
}
\description{
Runs \code{bmab_gi()} multiple times for a range of arm starting states. The starting states used are
determined by the arguments \code{alpha_start}, \code{beta_start}, and \code{num_actions} and are:
\itemize{
\item \code{alpha = alpha_start : (alpha_start_start + num_actions - 1)},
\item \code{beta = beta_start : (beta_start + num_actions - 1)},
\item where \code{alpha + beta <= num_actions + alpha_start + beta_start - 1}.
}

Alternatively, you can supply \code{Sigma_start} and \code{n_start} instead of \code{alpha_start} and \code{beta_start}.
The function will behave in the same way using \code{alpha = Sigma} and  \code{beta = n - Sigma}. Give either
alpha/beta or Sigma/n - do not mix the two forms.
}
\examples{
bmab_gi_multiple(1, 1, gamma = 0.9, N = 80, num_actions = 5)
bmab_gi_multiple(gamma = 0.9, N = 80, num_actions = 5, Sigma_start = 1, n_start = 2)

}
