% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_multiple}
\alias{bmab_gi_multiple}
\title{Calculates Gittins indices for the BMAB with a range of arm states.}
\usage{
bmab_gi_multiple(Sigma_start = 1, n_start = 2, gamma, N, num_actions,
  tol)
}
\arguments{
\item{gamma}{numeric in (0, 1]; discount factor}

\item{N}{integer>0; time horizon used}

\item{num_actions}{determinines the number of states GI are calculated for}

\item{tol}{absolute accuracy required}

\item{Sigma_start=1}{lowest value of Sigma for the arms}

\item{n_start=2}{lowest value of n for the arms}
}
\value{
A triangular matrix of GI values
}
\description{
The states are a triangular matrix with:
Sigma = Sigma_start, Sigma_start + 1, ..., Sigma_start + num_actions - 1,
n = n_start, n_start + 1, ..., n_start + num_actions - 1,
and Sigma <= Sigma_start + n - n_start.
}
