% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_multiple}
\alias{bmab_gi_multiple}
\alias{bmab_gi_multiple_ab}
\alias{bmab_gi_ab}
\title{Calculate Gittins indices for multiple arms (Bernoulli rewards)}
\usage{
bmab_gi_multiple(Sigma_start = 1, n_start = 2, gamma, N, num_actions,
  tol)

bmab_gi_multiple_ab(alpha_start = 1, beta_start = 1, gamma, N,
  num_actions, tol)

bmab_gi_ab(alpha, beta, gamma, tol, N, lb = NA, ub = NA, kgi = F,
  giplus = F)
}
\arguments{
\item{gamma}{numeric in (0, 1]; discount factor}

\item{N}{integer>0; time horizon used}

\item{num_actions}{determinines the number of states GI are calculated for}

\item{tol}{absolute accuracy required}

\item{alpha}{value of alpha for the arm}

\item{beta}{value of beta for the arm}

\item{Sigma_start=1}{lowest value of Sigma for the arms}

\item{n_start=2}{lowest value of n for the arms}

\item{alpha_start=1}{lowest value of alpha for the arms}

\item{beta_start=1}{lowest value of beta for the arms}
}
\value{
A triangular matrix of GI values
}
\description{
There are two versions which differ only in their parameterisation and in the layout of the output.
\code{bmab_gi_multiple()} uses \code{Sigma} and \code{n} parameters,
while \code{bmab_gi_multiple_ab()} uses \code{alpha = Sigma} and  \code{beta = n - Sigma}.
}
\details{
The states are a triangular matrix with:
\itemize{
\item \code{Sigma = Sigma_start : Sigma_start + num_actions - 1},
\item \code{n = n_start : n_start + num_actions - 1},
\item \code{Sigma <= Sigma_start + n - n_start}.
}
}
\examples{
bmab_gi_multiple(1, 2, gamma = 0.9, N = 80, num_actions = 20, tol = 5e-5)
bmab_gi_multiple_ab(1, 1, gamma = 0.9, N = 80, num_actions = 20, tol = 5e-5)

}
