% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_gi_value}
\alias{bmab_gi_value}
\title{Value approximation for the one-armed bandit with Normal rewards.}
\usage{
bmab_gi_value(lambda, Sigma, n, gamma, N)
}
\arguments{
\item{lambda}{reward from the known arm}

\item{Sigma}{value of Sigma for the unknown arm}

\item{n}{value of n for the unknown arm}

\item{gamma}{discount factor}

\item{N}{time horizon used in the dynamic programme}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Value approximation for the one-armed bandit with Normal rewards.
}
