% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmab.R
\name{bmab_giplus_value}
\alias{bmab_giplus_value}
\title{Value of one-armed bandit using GI+ (Bernoulli rewards)}
\usage{
bmab_giplus_value(lambda, Sigma, n, gamma)
}
\arguments{
\item{lambda}{reward from the known arm.}

\item{Sigma}{mean of reward belief for the unknown arm.}

\item{n}{value of n for the unknown arm.}

\item{gamma}{numeric in (0, 1); discount factor.}
}
\value{
Difference in value between safe and unknown arms
}
\description{
Value of one-armed bandit using GI+ (Bernoulli rewards)
}
