% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{calibrate_arm}
\alias{calibrate_arm}
\title{Calibrate an unknown arm against a known arm}
\usage{
calibrate_arm(FUN, lb, ub, tol, ...)
}
\arguments{
\item{FUN}{a value function}

\item{lb}{initial lower end of interval}

\item{ub}{initial upper end of interval}

\item{tol}{absolute accuracy required}
}
\value{
A vector of lower and upper end of an interval containing the true value
}
\description{
Performs iterative reduction of an interval using calibration with value function FUN.
}
