% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gittins-package.R
\docType{package}
\name{gittins-package}
\alias{gittins}
\alias{gittins-package}
\title{gittins: Calculate Gittins Indices}
\description{
This is an R package to calculate Gittins indices for the multi-armed bandit problem.

This project contains functions written in R to calculate Gittins indices for
the Bayesian multi-armed bandit problem with Bernoulli or Normal rewards. Detailed information
on the methodology as well as the effect of different computation parameters on speed and
accuracy can be found in the paper at \url{https://arxiv.org/abs/1909.05075}
and in my thesis at \url{http://eprints.lancs.ac.uk/84589/} (in Chapter 5).

I welcome any feedback or suggestions (see link to Github page below). If the package is useful
in your work then can you please reference the package and/or the paper.
}
\seealso{
\itemize{
\item The main package Github page is \url{https://github.com/jedwards24/gittins}.
\item Report bugs at \url{https://github.com/jedwards24/gittins/issues}.
}
}
\author{
James Edwards (also maintainer)
}
