% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_args}
\alias{nmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{Numeric. Value of Sigma for the arm.}

\item{n}{Numeric > 0. Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tau}{Numeric > 0. Observation precision.}

\item{tol}{Numeric > 0. Absolute accuracy required.}

\item{N}{Integer >= 2. Time horizon used.}

\item{xi}{Numeric > 0. Value of xi (entent of dynamic program state space).}

\item{delta}{Numeric > 0. Value of delta (fineness of discretisation in the dynamic program).}
}
\description{
Function arguments
}
\keyword{internal}
