% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_args}
\alias{nmab_args}
\title{Function arguments}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}

\item{tau}{observation precision}

\item{tol}{absolute accuracy required}

\item{N}{time horizon used}

\item{xi}{value of xi (entent of dynamic programmme state space)}

\item{delta}{value of delta (fineness of discretisation in the dynamic programmme)}
}
\description{
Function arguments
}
\keyword{integral}
