% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi}
\alias{nmab_gi}
\title{Calculate the Gittins index for a single arm (normal rewards)}
\usage{
nmab_gi(Sigma, n, gamma, tau, tol, N, xi, delta, lb = NA, ub = NA)
}
\arguments{
\item{Sigma}{value of Sigma for the arm.}

\item{n}{value of n for the arm.}

\item{gamma}{numeric in (0, 1); discount factor.}

\item{tau}{observation precision.}

\item{tol}{absolute accuracy required.}

\item{N}{time horizon used.}

\item{xi}{value of xi (entent of dynamic programmme state space).}

\item{delta}{value of delta (fineness of discretisation in the dynamic programmme).}

\item{lb}{Optional lower bound for GI.}

\item{ub}{Optional upper bound for GI.}
}
\value{
A vector of GI values.
}
\description{
The initial interval for calibration are as follows:
For lower bound, use \code{lb} if supplied else use KGI.
For upper bound, use \code{ub} if supplied else use GI+.
}
