% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi}
\alias{nmab_gi}
\title{Calculates the Gittins index for the NMAB for a single arm with supplied parameters.}
\usage{
nmab_gi(Sigma, n, gamma, tau, tol, N, xi, delta, lb = NA, ub = NA,
  kgi = T, giplus = T)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}

\item{tau}{observation precision}

\item{tol}{absolute accuracy required}

\item{N}{time horizon used}

\item{xi}{value of xi (entent of dynamic programmme state space)}

\item{delta}{value of delta (fineness of discretisation in the dynamic programmme)}

\item{lb=NA}{optional lower bound for GI}

\item{ub=NA}{optional upper bound for GI}

\item{kgi=T}{optional boolean indicates whether to use KGI for lower bound (only if lb=NA)}

\item{giplus=T}{optional boolean indicates whether to use GI+ for upper bound (only if ub=NA)}
}
\value{
A vector of GI values
}
\description{
The initial interval for calibration are as follows:
For lower bound, use lb if supplied else use KGI if kgi=T or Sigma/n otherwise.
For upper bound, use ub if supplied else use GI+ if giplus=T or gamma / ((1 - gamma) * sqrt(n)) otherwise.
}
