% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi}
\alias{nmab_gi}
\title{Calculate the Gittins index for a single arm (normal rewards)}
\usage{
nmab_gi(Sigma, n, gamma, tau, N, xi, delta, tol = 5e-04, lb = NA, ub = NA)
}
\arguments{
\item{Sigma}{Value of Sigma for the arm.}

\item{n}{Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tau}{Observation precision.}

\item{N}{Time horizon used.}

\item{xi}{Value of xi (entent of dynamic program state space).}

\item{delta}{Value of delta (fineness of discretisation in the dynamic program).}

\item{tol}{Absolute accuracy required.}

\item{lb}{Optional lower bound for GI.}

\item{ub}{Optional upper bound for GI.}
}
\value{
A vector of GI values.
}
\description{
The initial interval for calibration are as follows:
\itemize{
\item For lower bound, use \code{lb} if supplied else use KGI.
\item For upper bound, use \code{ub} if supplied else use GI+.
}
}
