% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi_multiple}
\alias{nmab_gi_multiple}
\title{Calculate Gittins indices for multiple arms (normal rewards)}
\usage{
nmab_gi_multiple(n_range, gamma, tau, N, xi, delta, tol = 5e-04)
}
\arguments{
\item{n_range}{Numeric vector giving values of n (all greater than 0).}

\item{gamma}{Numeric in (0, 1). Reward discount factor.}

\item{tau}{Numeric > 0. Observation precision.}

\item{N}{Integer >= 2. Time horizon used.}

\item{xi}{Numeric > 0. Value of xi (entent of dynamic program state space).}

\item{delta}{Numeric > 0. Value of delta (fineness of discretisation in the dynamic program).}

\item{tol}{Numeric > 0. Absolute accuracy required.}
}
\value{
A data frame of GI vales with a row for each n in \code{n_range}. The parameters used are
attached as an attribute \code{params}.
}
\description{
This assumes mu = 0 as the GI for other values can then be derived by simple addition. See
\code{?nmab_gi} for more detail on the problem and computation parameters.
}
\examples{
n1 <- nmab_gi_multiple(1:4, gamma = 0.9, tau = 1, N = 30, xi = 3, delta = 0.02, tol = 5e-4)
n1
attr(n1, "params")
}
