% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_gi_multiple}
\alias{nmab_gi_multiple}
\title{Calculate the Gittins index for multiple arms (normal rewards)}
\usage{
nmab_gi_multiple(n_range, gamma, tau, N, xi, delta, tol = 0.0005)
}
\arguments{
\item{n_range}{Numeric vector giving values of n (must be ascending).}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tau}{Observation precision.}

\item{N}{Time horizon used.}

\item{xi}{Value of xi (entent of dynamic program state space).}

\item{delta}{Value of delta (fineness of discretisation in the dynamic program).}

\item{tol}{Absolute accuracy required.}
}
\value{
A data frame of GI vales with a row for each n in \code{n_range}. The parameters used are
attached as an attribute \code{params}.
}
\description{
Assumes mu = 0.
}
\examples{
n1 <- nmab_gi_multiple(1:4, gamma = 0.9, tau = 1, N = 30, xi = 3, delta = 0.02, tol = 5e-4)
n1
attr(n1, "params")
}
