% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_giplus}
\alias{nmab_giplus}
\title{Calculate the GI+ index for a single arm (normal rewards)}
\usage{
nmab_giplus(Sigma, n, gamma, tol = 0.0005, ub = NA, upper = FALSE)
}
\arguments{
\item{Sigma}{Value of Sigma for the arm.}

\item{n}{Value of n for the arm.}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tol}{Absolute accuracy required.}

\item{ub}{Optional upper bound for the index.}

\item{upper}{If TRUE, the upper end of the interval is returned, otherwise the midpoint.}
}
\value{
A vector of GI values.
}
\description{
The index is an upper bound for GI.
}
