% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_giplus_value}
\alias{nmab_giplus_value}
\title{Value of one-armed bandit using GI+ (normal rewards)}
\usage{
nmab_giplus_value(lambda, mu, n, gamma)
}
\arguments{
\item{lambda}{Reward from the known arm}

\item{mu}{Mean of reward belief for the unknown arm}

\item{n}{Value of n for the unknown arm}

\item{gamma}{Numeric in (0, 1); discount factor.}
}
\value{
Difference in value between safe and unknown arms.
}
\description{
Value of one-armed bandit using GI+ (normal rewards)
}
