% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_kgi}
\alias{nmab_kgi}
\title{Calculate the knowledge gradient index for a single arm (normal rewards)}
\usage{
nmab_kgi(Sigma, n, gamma, tau, tol = 5e-04, ub = NA, lower = FALSE)
}
\arguments{
\item{Sigma}{Numeric. Value of Sigma for the arm (Bayesian reward).}

\item{n}{Numeric > 0. Value of n for the arm (Bayesian number of observation).}

\item{gamma}{Numeric in (0, 1). Reward discount factor.}

\item{tau}{Numeric > 0. Observation precision.}

\item{tol}{Numeric > 0. Absolute accuracy required.}

\item{ub}{Optional upper bound for the index.}

\item{lower}{If TRUE, the lower end of the interval is returned, otherwise the midpoint.}
}
\value{
A vector of GI values.
}
\description{
The index is an lower bound for GI.
}
