% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_kgi}
\alias{nmab_kgi}
\title{Calculates the KGI index for the NMAB for a single arm with supplied parameters.}
\usage{
nmab_kgi(Sigma, n, gamma, tau, tol, ub, lower = F)
}
\arguments{
\item{Sigma}{value of Sigma for the arm}

\item{n}{value of n for the arm}

\item{gamma}{discount factor}

\item{tau}{observation precision}

\item{tol}{absolute accuracy required}

\item{lower=F}{if TRUE, the lower end of the interval is returned, otherwise the midpoint}
}
\value{
A vector of GI values
}
\description{
The index is an lower bound for GI.
}
