% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_kgi_value}
\alias{nmab_kgi_value}
\title{Value of one-armed bandit using KGI (normal rewards)}
\usage{
nmab_kgi_value(lambda, mu, n, gamma, tau)
}
\arguments{
\item{lambda}{Reward from the known arm}

\item{mu}{mMan of reward belief for the unknown arm}

\item{n}{Value of n for the unknown arm}

\item{gamma}{Numeric in (0, 1); discount factor.}

\item{tau}{Observation precision.}
}
\value{
Difference in value between safe and unknown arms.
}
\description{
Calculation is as given in Ryzhov, Powell & Frazier (2012), simplified since \code{lambda >= mu}
always when called here.
}
