% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_risky_reward}
\alias{nmab_risky_reward}
\title{Reward of the risky arm in a one-armed bandit process}
\usage{
nmab_risky_reward(mu, y_lo_scaled, y_hi_scaled, tn_scaled, tau, s,
  value_vec, discount)
}
\description{
Helper function only used in nmab_gi_value.
}
\keyword{internal}
