% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmab.R
\name{nmab_v_args}
\alias{nmab_v_args}
\title{Function arguments for value functions}
\arguments{
\item{lambda}{Reward from the known arm}

\item{mu}{Mean of reward belief for the unknown arm}

\item{n}{Numeric > 0. Value of n for the unknown arm}
}
\description{
Function arguments for value functions
}
\keyword{internal}
